server/worker/top_5_budget/requirements.txt
aio-pika>=6.8.0
pydantic==2.0.3
python-dotenv==1.0.0